Comp/GithubCopilotAgent/sample/workspace/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
mysqlclient
python-dotenv
